Move marker factory funcs to new mod
